Deployment/requirements.txt
pandas==2.2.2
matplotlib==3.9.2
streamlit==1.37.1
plotly==5.24.1
scikit-learn==1.5.2
seaborn==0.13.2
numpy==1.26.4
tensorflow==2.10.1
pillow==9.5.0